Back to working state
